+++ packages/s2-react/src/hooks/usePagination.ts
@@ -1,5 +1,6 @@
 import React from 'react';
 import { S2Event, SpreadSheet } from '@antv/s2';
+import { useLatest } from 'ahooks';
 import { isEmpty } from 'lodash';
 import { BaseSheetComponentProps } from '../components';
 
@@ -10,8 +11,11 @@ export const usePagination = (
   s2: SpreadSheet,
   props: BaseSheetComponentProps,
 ) => {
-  const { options, dataCfg } = props;
-  const [total, setTotal] = React.useState<number>(0);
+  const { options } = props;
+  const [total, setTotal] = React.useState<number>(
+    s2?.facet?.viewCellHeights.getTotalLength() ?? 0,
+  );
+  const paginationRef = useLatest(options.pagination);
   const [current, setCurrent] = React.useState<number>(
     options.pagination?.current || DEFAULT_PAGE_NUMBER,
   );
@@ -19,10 +23,12 @@ export const usePagination = (
     options.pagination?.pageSize || DEFAULT_PAGE_SIZE,
   );
 
+  // sync state.pagination -> s2.pagination
   React.useEffect(() => {
-    if (!s2 || isEmpty(options.pagination)) {
+    if (!s2 || isEmpty(paginationRef.current)) {
       return;
     }
+
     s2.setOptions({
       pagination: {
         current,
@@ -30,15 +36,18 @@ export const usePagination = (
       },
     });
     s2.render(false);
-  }, [pageSize, current, options.pagination, s2]);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [pageSize, current, s2]);
 
+  // sync props.pagination -> state.pagination
   React.useEffect(() => {
     setCurrent(options?.pagination?.current || DEFAULT_PAGE_NUMBER);
     setPageSize(options?.pagination?.pageSize || DEFAULT_PAGE_SIZE);
   }, [options.pagination]);
 
+  // sync layout result total -> state.total
   React.useEffect(() => {
-    if (!s2 || isEmpty(options.pagination)) {
+    if (!s2 || isEmpty(paginationRef.current)) {
       return;
     }
 
@@ -47,7 +56,8 @@ export const usePagination = (
     return () => {
       s2.off(S2Event.LAYOUT_PAGINATION, totalUpdateCallback);
     };
-  }, [options.pagination, dataCfg, s2]);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [s2]);
 
   return {
     total,